Separate RidgeRegression class as Ridge.py file.
