+++ src/symfc_alm/__init__.py
@@ -2,7 +2,5 @@
 from .symfc_alm import CellDataset  # noqa: F401
 from .symfc_alm import DispForceDataset  # noqa: F401
 from .symfc_alm import LinearModel  # noqa: F401
-from .symfc_alm import RidgeRegression  # noqa: F401
 from .symfc_alm import SymfcAlm  # noqa: F401
 from .symfc_alm import read_dataset  # noqa: F401
-from .symfc_alm import standardize_data  # noqa: F401
